+++ cucumber-jparest/src/test/java/uk/gov/homeoffice/digital/sas/cucumberjparest/RunCucumberIntegrationTest.java
@@ -13,7 +13,7 @@ import org.junit.platform.suite.api.Suite;
 @IncludeEngines("cucumber")
 @SelectClasspathResource("uk/gov/homeoffice/digital/sas/cucumberjparest")
 @SelectClasspathResource("test")
-@ConfigurationParameter(key = GLUE_PROPERTY_NAME, value = "uk.gov.homeoffice.digital.sas.cucumberjparest,uk.gov.homeoffice.digital.sas.cucumberjparesttestapi")
+@ConfigurationParameter(key = GLUE_PROPERTY_NAME, value = "uk.gov.homeoffice.digital.sas.cucumberjparest,uk.gov.homeoffice.digital.sas.cucumberjparest.testapi")
 @ConfigurationParameter(key = PLUGIN_PROPERTY_NAME, value = "pretty,html:target/cucumber-report.html")
 @ConfigurationParameter(key = OBJECT_FACTORY_PROPERTY_NAME, value = "io.cucumber.spring.SpringFactory")
 @SuppressWarnings("squid:S2187")//Actual tests are BDD features under resources folder
